ft-SPLICE_SLICE 3 : Implemented Splice and Improved Slice
Wrote basic iteration of splice and is briefly tested. Renamed splice and overloads to slice_to and splice_to respectively. Removed redundant code in splice_to and slice_to. Not used remove_head() and remove_tail() directly into splice for the sake of readability of code in learn namespace. Better version will be written in CADS namespace in future commits