Main 0 : Project Reset
Making significant restructuring changes in the project. Establishing proper naming conventions for each commit and creating individual branches for each module of the project to run the development in parallel. In the end, the main.cpp will only contain sections of commented code that can be uncommented to run unit test of any particular data structure or algorithm in consideration. No conflicts will be there. Amen to this idea. My god.